frontend/requirements.txt
flask==3.0.3
numpy==1.24.4
opencv-python==4.9.0.80
Pillow==7.0.0
requests==2.22.0
